split catalogue api into 3 modules
